backend/requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
djoser
drf-spectacular
hyperlink
idna
incremental
inflection
jsonschema
jsonschema-specifications
msgpack
oauthlib
pycparser
PyJWT
pyotp
python-dotenv
python3-openid
PyYAML
qrcode
redis
referencing
requests
requests-oauthlib
rpds-py
social-auth-app-django
social-auth-core
sqlparse
tomli
twisted
txaio
typing-extensions
uritemplate
urllib3
gunicorn
daphne
hvac
django-redis
channels[daphne]
channels-redis
channels-binding
Pillow
psycopg2-binary
